feat: implement create guarantor functionality and refactor models (#2660)
* Move guarantor related models and DTOs to a dedicated `guarantor` sub-package for better modularity.
* Implement `CreatedGuarantor` domain model and `CreateGuarantorResponseDto` to handle guarantor creation responses.
* Update `LoanAccountSummaryRepository` and `CreateGuarantorUseCase` to return `CreatedGuarantor` instead of a generic response.
* Refactor `CreateGuarantorViewModel` to use renamed internal methods (`loadGuarantorForm`, `submitGuarantor`) and updated domain models.
* Prevent back navigation in `CreateGuarantorScreen` while a guarantor submission is in progress.
* Remove unused guarantor navigation events and actions from `LoanAccountSummaryViewModel` and related screen routes.
