+++ core/data/src/commonMain/kotlin/com/mifos/core/data/repositoryImp/LoanAccountSummaryRepositoryImp.kt
@@ -16,9 +16,9 @@ import com.mifos.core.data.mapper.loan.toDto
 import com.mifos.core.data.repository.LoanAccountSummaryRepository
 import com.mifos.core.data.util.NetworkMonitor
 import com.mifos.core.data.util.runAsDataState
-import com.mifos.core.model.entity.accounts.loan.GuarantorTemplate
-import com.mifos.core.model.objects.account.loan.CreateGuarantorInput
-import com.mifos.core.network.GenericResponse
+import com.mifos.core.model.objects.account.loan.guarantor.CreateGuarantorInput
+import com.mifos.core.model.objects.account.loan.guarantor.CreatedGuarantor
+import com.mifos.core.model.objects.account.loan.guarantor.GuarantorTemplate
 import com.mifos.core.network.datamanager.DataManagerLoan
 import com.mifos.room.entities.accounts.loans.LoanWithAssociationsEntity
 import kotlinx.coroutines.CoroutineDispatcher
@@ -52,12 +52,12 @@ class LoanAccountSummaryRepositoryImp(
     override suspend fun createGuarantor(
         loanId: Int,
         createGuarantorInput: CreateGuarantorInput,
-    ): DataState<GenericResponse> {
+    ): DataState<CreatedGuarantor> {
         return runAsDataState(
             networkMonitor,
             ioDispatcher,
         ) {
-            dataManagerLoan.createGuarantor(loanId, createGuarantorInput.toDto())
+            dataManagerLoan.createGuarantor(loanId, createGuarantorInput.toDto()).toDomain()
         }
     }
 }
