+++ core/model/src/commonMain/kotlin/com/mifos/core/model/objects/account/loan/guarantor/CreateGuarantorInput.kt
@@ -7,12 +7,9 @@
  *
  * See https://github.com/openMF/mifos-x-field-officer-app/blob/master/LICENSE.md
  */
-package com.mifos.core.network.model.guarantor
+package com.mifos.core.model.objects.account.loan.guarantor
 
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CreateGuarantorRequestDto(
+data class CreateGuarantorInput(
     val existingClientId: Int? = null,
     val clientRelationshipTypeId: Int,
     val firstname: String? = null,
