+++ core/model/src/commonMain/kotlin/com/mifos/core/model/objects/account/loan/guarantor/CreatedGuarantor.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright 2026 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mifos-x-field-officer-app/blob/master/LICENSE.md
+ */
+package com.mifos.core.model.objects.account.loan.guarantor
+
+data class CreatedGuarantor(
+    val resourceId: Long?,
+)
