+++ core/network/src/commonMain/kotlin/com/mifos/core/network/datamanager/DataManagerLoan.kt
@@ -24,8 +24,9 @@ import com.mifos.core.model.objects.account.loan.transfer.AccountTransferTemplat
 import com.mifos.core.network.BaseApiManager
 import com.mifos.core.network.GenericResponse
 import com.mifos.core.network.model.LoansPayload
-import com.mifos.core.network.model.guarantor.CreateGuarantorRequestDto
-import com.mifos.core.network.model.guarantor.GuarantorTemplateDto
+import com.mifos.core.model.objects.account.loan.guarantor.CreateGuarantorResponseDto
+import com.mifos.core.model.objects.account.loan.guarantor.GuarantorRequestDto
+import com.mifos.core.model.objects.account.loan.guarantor.GuarantorTemplateDto
 import com.mifos.room.entities.PaymentTypeOptionEntity
 import com.mifos.room.entities.accounts.loans.LoanRepaymentRequestEntity
 import com.mifos.room.entities.accounts.loans.LoanRepaymentResponseEntity
@@ -120,8 +121,8 @@ class DataManagerLoan(
 
     suspend fun createGuarantor(
         loanId: Int,
-        request: CreateGuarantorRequestDto,
-    ): GenericResponse {
+        request: GuarantorRequestDto,
+    ): CreateGuarantorResponseDto {
         return mBaseApiManager.loanService.createGuarantor(loanId, request)
     }
