+++ feature/loan/src/commonMain/kotlin/com/mifos/feature/loan/loanAccountSummary/LoanAccountSummaryScreenRoute.kt
@@ -34,7 +34,6 @@ fun NavGraphBuilder.loanAccountSummary(
     approveLoan: (loanId: Int, loanWithAssociations: LoanWithAssociationsEntity) -> Unit,
     disburseLoan: (Int) -> Unit,
     onRepaymentClick: (LoanWithAssociationsEntity) -> Unit,
-    onCreateGuarantorClicked: (Int) -> Unit,
     navController: NavController,
 ) {
     composable<LoanAccountSummaryScreenRoute> {
@@ -47,7 +46,6 @@ fun NavGraphBuilder.loanAccountSummary(
             approveLoan = approveLoan,
             disburseLoan = disburseLoan,
             onRepaymentClick = onRepaymentClick,
-            onCreateGuarantorClicked = onCreateGuarantorClicked,
             onNavigateBack = onBackPressed,
             navController = navController,
         )
